Added the ability to delete notes

+++ app/src/main/java/ru/trimok/notes/domain/usecase/DeleteUserNoteUseCase.kt
@@ -4,7 +4,7 @@ import ru.trimok.notes.domain.repository.NoteRepository
 
 class DeleteUserNoteUseCase(private val noteRepository: NoteRepository) {
 
-    fun execute(noteId: Int): Boolean{
+    fun execute(noteId: Long): Boolean{
         return noteRepository.deleteNote(noteId)
     }
 }
